Support longer user IDs

+++ schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "votes" (
-  "user_id" integer NOT NULL,
+  "user_id" bigint NOT NULL,
   "screen_name" varchar(20) DEFAULT NULL,
   "vote" integer NOT NULL,
   "when" timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
